Fixed pagination bug with wrong page number after changind positionsPerPage
