+++ src/components/Pagination/Pagination.jsx
@@ -1,10 +1,15 @@
 // компонент с пагинацией
+import {useEffect} from 'react';
 import styles from './Pagination.module.css';
 import cx from 'classnames';
 import { v4 as generateUid } from 'uuid';
 
-export const Pagination = ({ itemsPerPage, totalItems, currentPage, setPage }) => {
+const Pagination = ({ itemsPerPage, totalItems, currentPage, setPage }) => {
   const pageButtons = [];
+
+  useEffect(() => {
+    setPage(1)
+  }, [itemsPerPage])
 
   for (let i = 1; i <= Math.ceil(totalItems / itemsPerPage); i++) {
     pageButtons.push(i)
@@ -55,3 +60,5 @@ export const Pagination = ({ itemsPerPage, totalItems, currentPage, setPage }) =
     </>
   )
 }
+
+export default Pagination;
